Suppression de warnings : remplacement de size_t en int dans les DataTraits.

+++ ui/ListDataWidgetDialog.h
@@ -97,9 +97,9 @@ public:
 
 	void updateTable(const value_type& v)
 	{
-		size_t nb = rowTrait::size(v);
+        int nb = rowTrait::size(v);
 
-		for(size_t i = 0; i < nb; ++i)
+        for(int i = 0; i < nb; ++i)
 		{
 			const row_type* row = rowTrait::get(v, i);
 			if(row)
